pyproject.toml
[tool.poetry]
name = "drf-home-work"
version = "0.1.0"
description = ""
authors = ["George Astakhov <goha1115@mail.ru>"]
readme = "README.md"
packages = [{include = "drf_home_work"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
black = "^24.4.2"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
